+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 
 - The settings are now displayed in tabs.
 - Detail page side navigation points are greyed out if there is no applicable content.
+- A spinner is not displayed
 
 ### Fixed